fixed endpoints, all automatic tests pass

+++ Shuttle-back/src/main/java/com/shuttle/location/dto/LocationDTO.java
@@ -2,8 +2,8 @@ package com.shuttle.location.dto;
 
 public class LocationDTO {
 	private String address;
-	private long latitude;
-	private long longitude;
+	private Double latitude;
+	private Double longitude;
 	
 	public String getAddress() {
 		return address;
@@ -11,16 +11,16 @@ public class LocationDTO {
 	public void setAddress(String adress) {
 		this.address = adress;
 	}
-	public long getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(long latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
-	public long getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(long longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
